Start proto handling
