+++ ProtoBufs/UpdateProtos.sh
@@ -29,10 +29,12 @@ flightComputer="../OnPod/FlightComputer/include/ProtoBuffer/"
 flightComputerSrc="../OnPod/FlightComputer/src/"
 
 PDS="../OnLand/PDS/"
+NODESIM="../Testing/NodeSimulator/Protobuf"
 
 # Force copy the files, replacing existing ones
 cp -f NodeTelem/cpp/*.pb.h $flightComputer
 cp -f NodeTelem/cpp/*.pb.cc $flightComputerSrc
 cp -f PodTelem/py/* $PDS
+cp -f PodTelem/py/* $NODESIM
 
 echo "Update Complete"
